+++ contacts/contacts_request_builder.go
@@ -29,7 +29,7 @@ type ContactsRequestBuilderGetQueryParameters struct {
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Show only the first n items
-    Top *int64 `uriparametername:"%24top"`
+    Top *int32 `uriparametername:"%24top"`
 }
 // ContactsRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type ContactsRequestBuilderGetRequestConfiguration struct {
